functions/requirements.txt
firebase_functions>=0.2.0
firebase-admin
google-cloud-aiplatform[adk,agent_engines]==1.98.0
google-adk>=1.4.2
gofannon
requests # For fetching Gofannon manifest
google-cloud-logging>=3.0.0
litellm>=1.72.0
PyPDF>=5.6.0
PyGithub